共通処理切り出し
