--- src/app/yokohama/yellow_man/module/models/package-info.java
@@ -1,4 +1,0 @@
-/**
- * DBモデルクラスを提供する。
- */
-package yokohama.yellow_man.module.models;
