+++ src/app/yokohama/yellow_man/sena/components/db/CompanySchedulesComponent.java
@@ -1,44 +1,12 @@
 package yokohama.yellow_man.sena.components.db;
 
-import java.util.List;
-
-import com.avaje.ebean.Ebean;
-
-import yokohama.yellow_man.common_tools.ListUtils;
-import yokohama.yellow_man.module.models.CompanySchedules;
-
 /**
  * 企業スケジュール（company_schedules）モデルの操作を行うコンポーネントクラス。
+ * <p>共通コンポーネント{@link yokohama.yellow_man.sena.core.components.db.CompanySchedulesComponent}を拡張する。
  *
  * @author yellow-man
  * @since 1.0
+ * @see yokohama.yellow_man.sena.core.components.db.CompanySchedulesComponent
  */
-public class CompanySchedulesComponent {
-
-	/**
-	 * 検索条件に銘柄コード（{@code stock_code}）を指定し、
-	 * 直近の決算発表日（{@code settlement_date}）の企業スケジュールを返す。
-	 *
-	 * @param stockCode 銘柄コード
-	 * @return 直近の決算発表日（{@code settlement_date}）の企業スケジュールを返す。
-	 * 			取得できない場合は、{@code null}を返す。
-	 * @since 1.0
-	 */
-	public static CompanySchedules getCompanySchedulesByStockCode(Integer stockCode) {
-		List<CompanySchedules> list =
-				Ebean.find(CompanySchedules.class)
-					.where()
-					.eq("delete_flg", false)
-					.eq("stock_code", stockCode)
-					.orderBy("settlement_date DESC")
-					.findList();
-
-		CompanySchedules ret = null;
-		if (!ListUtils.isEmpty(list)) {
-			ret = list.get(0);
-		}
-
-		return ret;
-	}
-
+public class CompanySchedulesComponent extends yokohama.yellow_man.sena.core.components.db.CompanySchedulesComponent {
 }
