+++ src/app/yokohama/yellow_man/sena/core/components/db/dao/StocksDao.java
@@ -1,4 +1,4 @@
-package yokohama.yellow_man.sena.components.db.dao;
+package yokohama.yellow_man.sena.core.components.db.dao;
 
 import java.util.Date;
 import java.util.List;
@@ -6,8 +6,8 @@ import java.util.List;
 import com.avaje.ebean.Ebean;
 import com.avaje.ebean.SqlUpdate;
 
-import yokohama.yellow_man.module.models.Stocks;
 import yokohama.yellow_man.sena.components.db.StocksComponent;
+import yokohama.yellow_man.sena.core.models.Stocks;
 
 /**
  * 銘柄（stocks）モデルのネイティブクエリによる操作クラス。
