+++ src/app/yokohama/yellow_man/sena/core/components/db/dao/package-info.java
@@ -1,4 +1,4 @@
 /**
  * ネイティブSQLを必要とするデータアクセス機能を提供する。
  */
-package yokohama.yellow_man.sena.components.db.dao;
+package yokohama.yellow_man.sena.core.components.db.dao;
