+++ src/app/yokohama/yellow_man/sena/core/components/db/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 参照するアプリケーション全体で利用するデータアクセスに関するコンポーネントを提供する。
+ */
+package yokohama.yellow_man.sena.core.components.db;
