+++ src/app/yokohama/yellow_man/sena/core/components/package-info.java
@@ -1,4 +1,4 @@
 /**
  * 参照するアプリケーション全体で利用するデータアクセス、ログ管理、等機能に関するコンポーネントを提供する。
  */
-package yokohama.yellow_man.module.components;
+package yokohama.yellow_man.sena.core.components;
