+++ src/app/yokohama/yellow_man/sena/core/definitions/package-info.java
@@ -1,4 +1,4 @@
 /**
  * 参照するアプリケーション全体で利用する定数定義を提供する。
  */
-package yokohama.yellow_man.module.definitions;
+package yokohama.yellow_man.sena.core.definitions;
