+++ src/app/yokohama/yellow_man/sena/core/models/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * 参照するアプリケーション全体で利用するDBモデルクラスを提供する。
+ */
+package yokohama.yellow_man.sena.core.models;
