+++ src/app/yokohama/yellow_man/sena/jobs/AppJob.java
@@ -2,7 +2,7 @@ package yokohama.yellow_man.sena.jobs;
 
 import java.util.List;
 
-import yokohama.yellow_man.module.components.AppLogger;
+import yokohama.yellow_man.sena.core.components.AppLogger;
 
 /**
  * バッチ処理の基底クラス。
