+++ src/app/yokohama/yellow_man/sena/jobs/ImportIndicators.java
@@ -6,13 +6,13 @@ import java.util.Random;
 
 import play.Play;
 import yokohama.yellow_man.common_tools.ListUtils;
-import yokohama.yellow_man.module.components.AppLogger;
-import yokohama.yellow_man.module.models.Indicators;
-import yokohama.yellow_man.module.models.Stocks;
 import yokohama.yellow_man.sena.components.db.StocksComponent;
 import yokohama.yellow_man.sena.components.scraping.ScrapingComponent;
 import yokohama.yellow_man.sena.components.scraping.ScrapingException;
 import yokohama.yellow_man.sena.components.scraping.entity.IndicatorsEntity;
+import yokohama.yellow_man.sena.core.components.AppLogger;
+import yokohama.yellow_man.sena.core.models.Indicators;
+import yokohama.yellow_man.sena.core.models.Stocks;
 
 /**
  * 企業指標インポートバッチクラス。
